pyproject.toml
[project]
name = "aurinko-auth-server"
version = "0.1.0"
description = "OAuth server for Gmail authentication using Aurinko"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.100.0",
    "hypercorn>=0.14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
]
